requirements.txt
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
idna==2.10
install==1.3.4
isort==5.7.0
joblib==1.0.1
lazy-object-proxy==1.4.3
mccabe==0.6.1
nltk==3.6.2
oauthlib==3.1.0
pylint==2.6.2
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.0
requests-oauthlib==1.3.0
six==1.15.0
solace-pubsubplus==1.0.0
textblob==0.15.3
toml==0.10.2
tqdm==4.60.0
twitter-stream.py==0.8.4
typed-ast==1.4.2
urllib3==1.26.4
wrapt==1.12.1
